feat: created qb bridge
