+++ bridge/qb/client/events.lua
@@ -1,0 +1,5 @@
+-- Trigger Command
+--- @deprecated
+RegisterNetEvent('QBCore:Command:CallCommand', function(command)
+    ExecuteCommand(command)
+end)
