+++ bridge/qb/server/debug.lua
@@ -1,0 +1,6 @@
+---@deprecated Call lib.print.debug() instead
+---@param tbl any
+RegisterServerEvent('QBCore:DebugSomething', function(tbl)
+    local resource = GetInvokingResource() or "qbx-core"
+    lib.print.debug(resource, tbl)
+end)
