+++ bridge/qb/server/functions.lua
@@ -1,0 +1,140 @@
+local functions = require 'server.functions'
+
+---@deprecated
+functions.GetCoords = GetCoordsFromEntity
+
+---@deprecated use the native GetPlayerIdentifierByType?
+functions.GetIdentifier = GetPlayerIdentifierByType
+
+---@deprecated use the native GetPlayers instead
+functions.GetPlayers = GetPlayers
+
+---@deprecated Use functions.CreateVehicle instead.
+function functions.SpawnVehicle(source, model, coords, warp)
+    return SpawnVehicle(source, model, coords, warp)
+end
+
+---@deprecated use SpawnVehicle from imports/utils.lua
+functions.CreateVehicle = SpawnVehicle
+
+---@deprecated No replacement. See https://overextended.dev/ox_inventory/Functions/Client#useitem
+---@param source Source
+---@param item string name
+function functions.UseItem(source, item)
+    if GetResourceState('qb-inventory') == 'missing' then return end
+    exports['qb-inventory']:UseItem(source, item)
+end
+
+---@deprecated use KickWithReason from imports/utils.lua
+functions.Kick = KickWithReason
+
+---@deprecated use IsLicenseInUse from imports/utils.lua
+functions.IsLicenseInUse = IsLicenseInUse
+
+-- Utility functions
+
+---@deprecated use https://overextended.dev/ox_inventory/Functions/Server#search
+functions.HasItem = HasItem
+
+---@deprecated use GetPlate from imports/utils.lua
+functions.GetPlate = GetPlate
+
+-- Single add item
+---@deprecated incompatible with ox_inventory. Update ox_inventory item config instead.
+local function AddItem(itemName, item)
+    lib.print.warn(string.format("%s invoked Deprecated function AddItem. This is incompatible with ox_inventory", GetInvokingResource() or 'unknown resource'))
+    if type(itemName) ~= "string" then
+        return false, "invalid_item_name"
+    end
+
+    if QBCore.Shared.Items[itemName] then
+        return false, "item_exists"
+    end
+
+    QBCore.Shared.Items[itemName] = item
+
+    TriggerClientEvent('QBCore:Client:OnSharedUpdate', -1, 'Items', itemName, item)
+    TriggerEvent('QBCore:Server:UpdateObject')
+    return true, "success"
+end
+
+functions.AddItem = AddItem
+exports('AddItem', AddItem)
+
+-- Single update item
+---@deprecated incompatible with ox_inventory. Update ox_inventory item config instead.
+local function UpdateItem(itemName, item)
+    lib.print.warn(string.format("%s invoked deprecated function UpdateItem. This is incompatible with ox_inventory", GetInvokingResource() or 'unknown resource'))
+    if type(itemName) ~= "string" then
+        return false, "invalid_item_name"
+    end
+    if not QBCore.Shared.Items[itemName] then
+        return false, "item_not_exists"
+    end
+    QBCore.Shared.Items[itemName] = item
+    TriggerClientEvent('QBCore:Client:OnSharedUpdate', -1, 'Items', itemName, item)
+    TriggerEvent('QBCore:Server:UpdateObject')
+    return true, "success"
+end
+
+functions.UpdateItem = UpdateItem
+exports('UpdateItem', UpdateItem)
+
+-- Multiple Add Items
+---@deprecated incompatible with ox_inventory. Update ox_inventory item config instead.
+local function AddItems(items)
+    lib.print.warn(string.format("%s invoked deprecated function AddItems. This is incompatible with ox_inventory", GetInvokingResource() or 'unknown resource'))
+    local shouldContinue = true
+    local message = "success"
+    local errorItem = nil
+
+    for key, value in pairs(items) do
+        if type(key) ~= "string" then
+            message = "invalid_item_name"
+            shouldContinue = false
+            errorItem = items[key]
+            break
+        end
+
+        if QBCore.Shared.Items[key] then
+            message = "item_exists"
+            shouldContinue = false
+            errorItem = items[key]
+            break
+        end
+
+        QBCore.Shared.Items[key] = value
+    end
+
+    if not shouldContinue then return false, message, errorItem end
+    TriggerClientEvent('QBCore:Client:OnSharedUpdateMultiple', -1, 'Items', items)
+    TriggerEvent('QBCore:Server:UpdateObject')
+    return true, message, nil
+end
+
+functions.AddItems = AddItems
+exports('AddItems', AddItems)
+
+-- Single Remove Item
+---@deprecated incompatible with ox_inventory. Update ox_inventory item config instead.
+local function RemoveItem(itemName)
+    lib.print.warn(string.format("%s invoked deprecated function RemoveItem. This is incompatible with ox_inventory", GetInvokingResource() or 'unknown resource'))
+    if type(itemName) ~= "string" then
+        return false, "invalid_item_name"
+    end
+
+    if not QBCore.Shared.Items[itemName] then
+        return false, "item_not_exists"
+    end
+
+    QBCore.Shared.Items[itemName] = nil
+
+    TriggerClientEvent('QBCore:Client:OnSharedUpdate', -1, 'Items', itemName, nil)
+    TriggerEvent('QBCore:Server:UpdateObject')
+    return true, "success"
+end
+
+functions.RemoveItem = RemoveItem
+exports('RemoveItem', RemoveItem)
+
+return functions
