+++ bridge/qb/server/player.lua
@@ -1,0 +1,43 @@
+local playerObj = require 'server.player'
+
+---@deprecated ox_inventory automatically saves
+---@param source Source
+function playerObj.SaveInventory(source)
+    if GetResourceState('qb-inventory') == 'missing' then return end
+    exports['qb-inventory']:SaveInventory(source, false)
+end
+
+---@deprecated ox_inventory automatically saves
+---@param playerData PlayerData
+function playerObj.SaveOfflineInventory(playerData)
+    if GetResourceState('qb-inventory') == 'missing' then return end
+    exports['qb-inventory']:SaveInventory(playerData, true)
+end
+
+---@deprecated call ox_inventory exports directly
+---@param items any[]
+---@return number?
+function playerObj.GetTotalWeight(items)
+    if GetResourceState('qb-inventory') == 'missing' then return end
+    return exports['qb-inventory']:GetTotalWeight(items)
+end
+
+---@deprecated call ox_inventory exports directly
+---@param items any[]
+---@param itemName string
+---@return integer[]? slots
+function playerObj.GetSlotsByItem(items, itemName)
+    if GetResourceState('qb-inventory') == 'missing' then return end
+    return exports['qb-inventory']:GetSlotsByItem(items, itemName)
+end
+
+---@deprecated call ox_inventory exports directly
+---@param items any[]
+---@param itemName string
+---@return integer? slot
+function playerObj.GetFirstSlotByItem(items, itemName)
+    if GetResourceState('qb-inventory') == 'missing' then return end
+    return exports['qb-inventory']:GetFirstSlotByItem(items, itemName)
+end
+
+return playerObj
